[markup]: Implement new API
This is most likely the 1.0.0 API already, the previous version was
always temporary. The new API also makes the system more modular, and
thus allows us to cache in a more stable manner.

Performance isn't quite as good as it could be (there's likely a 2x
improvement to be found around the pre-processing step), but all things
in due time :)